Fix OtoL() to accumulate data in a long instead of an int
On 64-bit architectures where the long is bigger than the int, it makes
it possible to have files bigger than 2GB.

+++ lib/dpkg/tarfn.c
@@ -65,7 +65,7 @@ struct tar_header {
 static long
 OtoL(const char *s, int size)
 {
-	int n = 0;
+	long n = 0;
 
 	while (*s == ' ') {
 		s++;
